Fixed some bugs with geckodriver, implemented display tests
Gecko driver seems to have problems with accepting commands outside of start and quit. I'm not sure if this has to do with the driver itself (since it's version 0.16.1) or if it has to do with the library I'm using. I'll have to look into it. Tests are now restructured to allow for plugging in multiple different browsers, but since Gecko is broken for some reason, that functionality isn't implemented just yet.