Gutted email input verification, modularized month tests
I realized I was checking the wrong field for reminder e-mails. Since this is kept in the Git history, it can be referred to if  needed to check every shoe result for an e-mail input text box. It's not necessary according to the requirements at the moment, so I gutted the method to replace it with the reminder e-mails check.