Commented out empty collection check, implemented reminder e-mail check
The empty collection check was causing failures to whenever there were no results shown on the page (December has no shoe results). The current algorithm already checks if there are blank values when they are pulled in. Checking for an empty collection just means that there were no results returned, not that values are missing. This condition may be changed in the future, but at the moment, it's just causing false positives.